apple_captive_test {pingr}	R Documentation

Download Apple's captive portal test

Description

If the test page, returns "Success" that means that the computer is connected to the Internet.

Usage

apple_captive_test()

Details

Note that this function will fail if the computer is offline. Use is_online() to check if the computer is online.

Examples

apple_captive_test()

[Package pingr version 2.0.1 Index]