ping {pingr}	R Documentation

Ping a remote server, to see if it is alive

Description

This is the classic ping, using ICMP packages. Only the system administrator can send ICMP packages, so we call out to the system's ping utility.

Usage

ping(destination, continuous = FALSE, verbose = continuous, count = 3L,
  timeout = 1)

Arguments

destination
Host name or IP address.

continuous
Logical, whether to keep pinging until the user interrupts.

verbose
Whether to print progress on the screen while pinging.

count
Number of pings to perform.

timeout
Timeout for a ping response.

Value

Vector of response times. NA means no response, in seconds. Currently NAs are always at the end of the vector, and not in their correct position.

[Package pingr version 1.1.2 Index]