pingr-package {pingr}	R Documentation

Check if the local or remote computer is up

Description

Check if a remote computer is up. It can either just call the system ping command, or check a specified TCP port.

Author(s)

Maintainer: Gábor Csárdi csardi.gabor@gmail.com

See Also

Useful links:

https://github.com/r-lib/pingr#readme

Report bugs at https://github.com/r-lib/pingr/issues

[Package pingr version 2.0.1 Index]