ping_port {pingr}	R Documentation

Check if a port of a server is active, measure response time

Description

Check if a port of a server is active, measure response time

Usage

ping_port(destination, port = 80L, continuous = FALSE,
  verbose = continuous, count = 3L, timeout = 1)

Arguments

destination
Host name or IP address.

port
Port.

continuous
Logical, whether to keep pinging until the user interrupts.

verbose
Whether to print progress on the screen while pinging.

count
Number of pings to perform.

timeout
Timeout, in seconds. How long to wait for a ping to succeed.

Value

Vector of response times. NA means no response

[Package pingr version 1.1.2 Index]